Added cross-matrix wrapper for Matrix.Get<>(); Moved float- and push-related stuff to partial class
